Create chart and add styling/animation
Add styling/animation to chart for visual purposes.
However, when the screen starts shrinking horizontally,
only the chart will shrink and not its siblings.

According to the documentation in Charts.js, you
cannot use relative units in order for the chart
to shrink relative to its parent.

To address this, flexbox will have to be removed
and apply position: relative for all child elements
of the main container.

For now, it will remain unaddressed as it is not
a priority.