Modify each chart's positioning
Adjust positioning of each chart relative to its parent
container. For aesthetic purposes, the spacing of each
chart is different, thus the values of the CSS property:
margin-left are different. If the values of the CSS property:
margin-left were all the same, the charts's spacing would
have looked inconsistent. Therefore, each of the chart's
spacing had to be addressed individually.